requirements.txt
Pillow==9.0.0
opencv-python==4.5.1.48
depthai==2.13.3.0
blobconverter==1.2.8
openvino
